move item handlers to inside item class mixin + fix accidental package move
