+++ fabric-item-groups-v0/src/main/java/net/fabricmc/fabric/api/client/itemgroup/FabricItemGroupBuilder.java
@@ -25,7 +25,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.Identifier;
 
-import net.fabricmc.fabric.impl.client.item.group.ItemGroupExtensions;
+import net.fabricmc.fabric.impl.item.group.ItemGroupExtensions;
 
 public final class FabricItemGroupBuilder {
 	private Identifier identifier;
